Add files via upload
In this project, we will make a program for Controlling Speed of DC Motor in MATLAB by creating a GUI having different buttons on it to control the direction of rotation of the DC motor as well as the speed of the DC motor by sending different commands through the serial port. We will be integrating Arduino with MATLAB to get optimum results.